Clean up `trapframe` items
