+++ kernel/src/arch/x86/cpu.rs
@@ -7,7 +7,7 @@ use alloc::{
 };
 
 use ostd::{
-    cpu::context::{cpuid, CpuException, CpuExceptionInfo, RawGeneralRegs, UserContext},
+    cpu::context::{cpuid, CpuException, CpuExceptionInfo, GeneralRegs, UserContext},
     Pod,
 };
 
@@ -102,11 +102,11 @@ macro_rules! copy_gp_regs {
 }
 
 impl GpRegs {
-    pub fn copy_to_raw(&self, dst: &mut RawGeneralRegs) {
+    pub fn copy_to_raw(&self, dst: &mut GeneralRegs) {
         copy_gp_regs!(self, dst);
     }
 
-    pub fn copy_from_raw(&mut self, src: &RawGeneralRegs) {
+    pub fn copy_from_raw(&mut self, src: &GeneralRegs) {
         copy_gp_regs!(src, self);
     }
 }
