+++ ostd/src/arch/riscv/cpu/context.rs
@@ -6,16 +6,15 @@ use core::fmt::Debug;
 
 use riscv::register::scause::{Exception, Interrupt, Trap};
 
-pub use crate::arch::trap::GeneralRegs as RawGeneralRegs;
 use crate::{
     arch::{
         timer::handle_timer_interrupt,
-        trap::{TrapFrame, UserContext as RawUserContext},
+        trap::{RawUserContext, TrapFrame},
     },
     user::{ReturnReason, UserContextApi, UserContextApiInternal},
 };
 
-/// Cpu context, including both general-purpose registers and FPU state.
+/// Userspace CPU context, including both general-purpose registers and FPU state.
 #[derive(Clone, Copy, Debug)]
 #[repr(C)]
 pub struct UserContext {
@@ -23,6 +22,45 @@ pub struct UserContext {
     trap: Trap,
     fpu_state: FpuState,
     cpu_exception_info: CpuExceptionInfo,
+}
+
+/// General registers.
+#[derive(Debug, Default, Clone, Copy)]
+#[repr(C)]
+#[expect(missing_docs)]
+pub struct GeneralRegs {
+    pub zero: usize,
+    pub ra: usize,
+    pub sp: usize,
+    pub gp: usize,
+    pub tp: usize,
+    pub t0: usize,
+    pub t1: usize,
+    pub t2: usize,
+    pub s0: usize,
+    pub s1: usize,
+    pub a0: usize,
+    pub a1: usize,
+    pub a2: usize,
+    pub a3: usize,
+    pub a4: usize,
+    pub a5: usize,
+    pub a6: usize,
+    pub a7: usize,
+    pub s2: usize,
+    pub s3: usize,
+    pub s4: usize,
+    pub s5: usize,
+    pub s6: usize,
+    pub s7: usize,
+    pub s8: usize,
+    pub s9: usize,
+    pub s10: usize,
+    pub s11: usize,
+    pub t3: usize,
+    pub t4: usize,
+    pub t5: usize,
+    pub t6: usize,
 }
 
 /// CPU exception information.
@@ -66,12 +104,12 @@ impl CpuExceptionInfo {
 
 impl UserContext {
     /// Returns a reference to the general registers.
-    pub fn general_regs(&self) -> &RawGeneralRegs {
+    pub fn general_regs(&self) -> &GeneralRegs {
         &self.user_context.general
     }
 
     /// Returns a mutable reference to the general registers
-    pub fn general_regs_mut(&mut self) -> &mut RawGeneralRegs {
+    pub fn general_regs_mut(&mut self) -> &mut GeneralRegs {
         &mut self.user_context.general
     }
 
@@ -166,15 +204,15 @@ impl UserContextApi for UserContext {
     }
 
     fn set_instruction_pointer(&mut self, ip: usize) {
-        self.user_context.set_ip(ip);
+        self.user_context.sepc = ip;
     }
 
     fn stack_pointer(&self) -> usize {
-        self.user_context.get_sp()
+        self.sp()
     }
 
     fn set_stack_pointer(&mut self, sp: usize) {
-        self.user_context.set_sp(sp);
+        self.set_sp(sp);
     }
 }
