+++ ostd/src/arch/riscv/trap/trap.rs
@@ -16,7 +16,7 @@
 
 use core::arch::{asm, global_asm};
 
-use crate::Pod;
+use crate::arch::cpu::context::GeneralRegs;
 
 #[cfg(target_arch = "riscv32")]
 global_asm!(
@@ -86,136 +86,29 @@ pub struct TrapFrame {
 }
 
 /// Saved registers on a trap.
-#[derive(Debug, Default, Clone, Copy, Pod)]
+#[derive(Debug, Default, Clone, Copy)]
 #[repr(C)]
-pub struct UserContext {
+pub(in crate::arch) struct RawUserContext {
     /// General registers
-    pub general: GeneralRegs,
+    pub(in crate::arch) general: GeneralRegs,
     /// Supervisor Status
-    pub sstatus: usize,
+    pub(in crate::arch) sstatus: usize,
     /// Supervisor Exception Program Counter
-    pub sepc: usize,
+    pub(in crate::arch) sepc: usize,
 }
 
-impl UserContext {
-    /// Go to user space with the context, and come back when a trap occurs.
+impl RawUserContext {
+    /// Goes to user space with the context, and comes back when a trap occurs.
     ///
     /// On return, the context will be reset to the status before the trap.
-    /// Trap reason and error code will be returned.
-    ///
-    /// # Example
-    /// ```no_run
-    /// use trapframe::{UserContext, GeneralRegs};
-    ///
-    /// // init user space context
-    /// let mut context = UserContext {
-    ///     general: GeneralRegs {
-    ///         sp: 0x10000,
-    ///         ..Default::default()
-    ///     },
-    ///     sepc: 0x1000,
-    ///     ..Default::default()
-    /// };
-    /// // go to user
-    /// context.run();
-    /// // back from user
-    /// println!("back from user: {:#x?}", context);
-    /// ```
-    pub fn run(&mut self) {
+    /// Trap reason and error code will be placed at `scause` and `stval`.
+    pub(in crate::arch) fn run(&mut self) {
         unsafe { run_user(self) }
-    }
-}
-
-/// General registers
-#[derive(Debug, Default, Clone, Copy, Pod)]
-#[repr(C)]
-#[expect(missing_docs)]
-pub struct GeneralRegs {
-    pub zero: usize,
-    pub ra: usize,
-    pub sp: usize,
-    pub gp: usize,
-    pub tp: usize,
-    pub t0: usize,
-    pub t1: usize,
-    pub t2: usize,
-    pub s0: usize,
-    pub s1: usize,
-    pub a0: usize,
-    pub a1: usize,
-    pub a2: usize,
-    pub a3: usize,
-    pub a4: usize,
-    pub a5: usize,
-    pub a6: usize,
-    pub a7: usize,
-    pub s2: usize,
-    pub s3: usize,
-    pub s4: usize,
-    pub s5: usize,
-    pub s6: usize,
-    pub s7: usize,
-    pub s8: usize,
-    pub s9: usize,
-    pub s10: usize,
-    pub s11: usize,
-    pub t3: usize,
-    pub t4: usize,
-    pub t5: usize,
-    pub t6: usize,
-}
-
-impl UserContext {
-    /// Get number of syscall
-    pub fn get_syscall_num(&self) -> usize {
-        self.general.a7
-    }
-
-    /// Get return value of syscall
-    pub fn get_syscall_ret(&self) -> usize {
-        self.general.a0
-    }
-
-    /// Set return value of syscall
-    pub fn set_syscall_ret(&mut self, ret: usize) {
-        self.general.a0 = ret;
-    }
-
-    /// Get syscall args
-    pub fn get_syscall_args(&self) -> [usize; 6] {
-        [
-            self.general.a0,
-            self.general.a1,
-            self.general.a2,
-            self.general.a3,
-            self.general.a4,
-            self.general.a5,
-        ]
-    }
-
-    /// Set instruction pointer
-    pub fn set_ip(&mut self, ip: usize) {
-        self.sepc = ip;
-    }
-
-    /// Set stack pointer
-    pub fn set_sp(&mut self, sp: usize) {
-        self.general.sp = sp;
-    }
-
-    /// Get stack pointer
-    pub fn get_sp(&self) -> usize {
-        self.general.sp
-    }
-
-    /// Set tls pointer
-    pub fn set_tls(&mut self, tls: usize) {
-        self.general.gp = tls;
     }
 }
 
 #[expect(improper_ctypes)]
 extern "C" {
     fn trap_entry();
-    fn run_user(regs: &mut UserContext);
+    fn run_user(regs: &mut RawUserContext);
 }
