+++ ostd/src/arch/x86/cpu/context/mod.rs
@@ -22,7 +22,10 @@ use x86_64::registers::{
 };
 
 use crate::{
-    arch::CPU_FEATURES,
+    arch::{
+        trap::{RawUserContext, TrapFrame},
+        CPU_FEATURES,
+    },
     task::scheduler,
     trap::call_irq_callback_functions,
     user::{ReturnReason, UserContextApi, UserContextApiInternal},
@@ -38,17 +41,40 @@ cfg_if! {
 
 pub use x86::cpuid;
 
-pub use crate::arch::trap::{
-    GeneralRegs as RawGeneralRegs, TrapFrame, UserContext as RawUserContext,
-};
-
-/// Cpu context, including both general-purpose registers and FPU state.
+/// Userspace CPU context, including both general-purpose registers and FPU state.
 #[derive(Clone, Default, Debug)]
 #[repr(C)]
 pub struct UserContext {
     user_context: RawUserContext,
     fpu_state: FpuState,
     cpu_exception_info: CpuExceptionInfo,
+}
+
+/// General registers.
+#[derive(Debug, Default, Clone, Copy, Eq, PartialEq)]
+#[repr(C)]
+#[expect(missing_docs)]
+pub struct GeneralRegs {
+    pub rax: usize,
+    pub rbx: usize,
+    pub rcx: usize,
+    pub rdx: usize,
+    pub rsi: usize,
+    pub rdi: usize,
+    pub rbp: usize,
+    pub rsp: usize,
+    pub r8: usize,
+    pub r9: usize,
+    pub r10: usize,
+    pub r11: usize,
+    pub r12: usize,
+    pub r13: usize,
+    pub r14: usize,
+    pub r15: usize,
+    pub rip: usize,
+    pub rflags: usize,
+    pub fsbase: usize,
+    pub gsbase: usize,
 }
 
 /// CPU exception information.
@@ -65,12 +91,12 @@ pub struct CpuExceptionInfo {
 
 impl UserContext {
     /// Returns a reference to the general registers.
-    pub fn general_regs(&self) -> &RawGeneralRegs {
+    pub fn general_regs(&self) -> &GeneralRegs {
         &self.user_context.general
     }
 
     /// Returns a mutable reference to the general registers
-    pub fn general_regs_mut(&mut self) -> &mut RawGeneralRegs {
+    pub fn general_regs_mut(&mut self) -> &mut GeneralRegs {
         &mut self.user_context.general
     }
