+++ ostd/src/arch/x86/cpu/context/tdx.rs
@@ -4,7 +4,7 @@ use tdx_guest::{
     handle_virtual_exception as do_handle_virtual_exception, tdcall, TdgVeInfo, TdxTrapFrame,
 };
 
-use crate::cpu::context::{RawGeneralRegs, UserContext};
+use super::{GeneralRegs, UserContext};
 
 pub(crate) struct VirtualizationExceptionHandler {
     ve_info: TdgVeInfo,
@@ -34,7 +34,7 @@ impl VirtualizationExceptionHandler {
     }
 }
 
-struct GeneralRegsWrapper<'a>(&'a mut RawGeneralRegs);
+struct GeneralRegsWrapper<'a>(&'a mut GeneralRegs);
 
 impl TdxTrapFrame for GeneralRegsWrapper<'_> {
     fn rax(&self) -> usize {
